Moved user dashboard styling to stylesheets
